worker/requirements.txt
audioop-lts==0.2.1
av==13.1.0
beautifulsoup4==4.13.4
click==8.1.8
cloup==3.0.7
colorama==0.4.6
decorator==5.2.1
glcontext==3.0.0
isosurfaces==0.1.2
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
networkx==3.4.2
numpy==2.2.5
pillow==11.2.1
pycairo==1.28.0
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
rich==14.0.0
scipy==1.15.2
screeninfo==0.8.1
skia-pathops==0.8.0.post2
soupsieve==2.7
srt==3.5.3
svgelements==1.9.6
tqdm==4.67.1
typing_extensions==4.13.2
watchdog==6.0.0
fastapi
uvicorn
redis
rq
openai
supabase-py
psycopg2-binary
manim
